refactor runner pahse 4 (completed)
