+++ src/pages/simulationpage/SimulatorPage.jsx
@@ -6674,6 +6674,13 @@ export function SimulatorPage({ gamificationMode = false }) {
         const msg = event.data;
         const msgArrivalMs = performance.now();
 
+        if (msg.type === 'error') {
+          appendConsoleEntry('error', `[SIM] ${msg.message || 'Runner error'}`, 'simulator');
+          logSerial(`Runner error: ${msg.message || 'Unknown error'}`, 'var(--red)');
+          handleStop();
+          return;
+        }
+
         if (msg.type === 'debug' && msg.category === 'rp2040-runtime') {
           const incomingBoardId = String(msg.boardId || '').trim();
           const hasKnownBoard = incomingBoardId && boardComponents.some((b) => b.id === incomingBoardId);
